\name{dpqrcox-package}
\alias{dpqrcox-package}
\alias{dpqrcox}
\docType{package}
\title{
\packageTitle{dpqrcox}
}
\description{
\packageDescription{dpqrcox}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{dpqrcox}
\packageIndices{dpqrcox}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dpqrcox}
Qihong Duan, Junrong Liu
Maintainer: \packageMaintainer{dpqrcox}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
